requirements.txt
click
dnspython
requests
